+++ src/Components/Broker/AbstractInterfaces.py
@@ -3,18 +3,32 @@ import sys
 import inspect
 from abc import ABC, abstractmethod
 import datetime as dt
+import time
 
 currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
 parentdir = os.path.dirname(currentdir)
 sys.path.insert(0, parentdir)
 
 
-class AccountInterface(ABC):
-    # This MUST not be overwritten. Use the "initialise()" to init a children interface
+class AbstractInterface(ABC):
     def __init__(self, config):
         self._config = config
         self._last_call_ts = dt.datetime.now()
         self.initialise()
+
+    def _wait_before_call(self, timeout: float):
+        """
+        Wait between API calls to not overload the server
+        """
+        while (dt.datetime.now() - self._last_call_ts) <= dt.timedelta(seconds=timeout):
+            time.sleep(0.5)
+        self._last_call_ts = dt.datetime.now()
+
+
+class AccountInterface(AbstractInterface):
+    # This MUST not be overwritten. Use the "initialise()" to init a children interface
+    def __init__(self, config):
+        super().__init__(config)
 
     @abstractmethod
     def initialise(self):
@@ -70,20 +84,13 @@ class AccountInterface(ABC):
 
     # No need to override this
     def _wait_before_call(self, timeout: float):
-        """
-        Wait between API calls to not overload the server
-        """
-        while (dt.datetime.now() - self._last_call_ts) <= dt.timedelta(seconds=timeout):
-            time.sleep(0.5)
-        self._last_call_ts = dt.datetime.now()
+        super()._wait_before_call(timeout)
 
 
-class StocksInterface(ABC):
+class StocksInterface(AbstractInterface):
     # This MUST not be overwritten. Use the "initialise()" to init a children interface
     def __init__(self, config):
-        self._config = config
-        self._last_call_ts = dt.datetime.now()
-        self.initialise()
+        super().__init__(config)
 
     @abstractmethod
     def initialise(self):
@@ -99,9 +106,4 @@ class StocksInterface(ABC):
 
     # No need to override this
     def _wait_before_call(self, timeout: float):
-        """
-        Wait between API calls to not overload the server
-        """
-        while (dt.datetime.now() - self._last_call_ts) <= dt.timedelta(seconds=timeout):
-            time.sleep(0.5)
-        self._last_call_ts = dt.datetime.now()
+        super()._wait_before_call(timeout)