+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## []
+### Changed
+- Improved and expanded onfiguration file format
+
+### Fixed
+- Broker package missing __init__.py
+
 ### Removed
 - Support of Python 3.5